+++ packages/shared/src/experience-api.ts
@@ -1939,4 +1939,3 @@ export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDa
       }),
   };
 }
-/* eslint-enable */
