+++ packages/shared/src/index.ts
@@ -1,1 +1,2 @@
-export { Api } from './experience-api/index.js';
+export { Api } from './experience-api.js';
+export * from './utils.js';
