+++ packages/sign-in-with-password/index.html
@@ -11,8 +11,8 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div class="app">
-      <form id="sign-in-form">
-        <img class="logo" alt="Logto branding logo" />
+      <form>
+        <img fetchpriority="high" class="logo" src="https://logto.io/logo.svg" alt="Logto branding logo" />
         <div class="input-field">
           <label for="username">Username</label>
           <input type="text" name="username" required />
